[skip ci]feat: Change download lib to RxDownload
download utils: done, need clean code
UI: doing

+++ app/src/main/java/net/xzos/upgradeall/ui/filemanagement/FileHubListViewHolder.kt
@@ -30,7 +30,7 @@ class FileHubListViewHolder(private val binding: ItemHubFileTaskerBinding)
                 binding.tvDownloading.text = downloadingNum
                 binding.tvPaused.text = pausedNum
                 binding.tvFailed.text = failedNum
-                binding.pbDownload.progress = downloadProgress
+                binding.pbDownload.progress = downloadProgress.toInt()
             }
         }
     }
